pyproject.toml
[project]
name = "cosmo-rag"
version = "0.1.0"
description = "Multimodal RAG system with TechReport dataset"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "faiss-cpu", 
    "datasets",
    "sentence-transformers",
    "uvicorn",
    "pillow",
    "tqdm",
    "rich",
    "pyyaml",
    "numpy",
    "transformers"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ext", "agents", "retrieval", "models", "prompts"]

[tool.uv]
dev-dependencies = []